requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aioschedule==0.5.2
aiosignal==1.2.0
aiosqlite==0.17.0
arrow==1.2.3
async-timeout==3.0.1
asyncpg==0.23.0
asyncudp==0.4.0
attrs==21.2.0
Automat==20.2.0
base58==0.2.2
bencode.py==4.0.0
bipwallet==1.0.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.1
click==6.6
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
docker-py==1.8.0
ecdsa==0.18.0
flake8==5.0.4
frozenlist==1.3.1
funcsigs==1.0.2
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.10.1
jmespath==0.10.0
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==4.16.0
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.7.0
mnemonic==0.13
more-itertools==9.0.0
multidict==5.1.0
packaging==21.3
parsel==1.6.0
path==16.5.0
path.py==12.5.0
pbkdf2==1.3
pexpect==4.8.0
Pillow==9.2.0
pluggy==1.0.0
priority==1.3.0
Protego==0.1.16
protobuf==3.0.0a3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.20
pycryptodome==3.15.0
PyDispatcher==2.0.5
pyflakes==2.5.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
PyYAML==6.0
qrcode==7.3.1
queuelib==1.6.1
requests==2.11.1
sanic==21.3.4
Sanic-Auth==0.3.0
sanic-routing==0.6.2
sanic-session==0.8.0
schedule==1.1.0
Scrapy==2.5.0
service-identity==21.1.0
sha256==0.3
six==1.16.0
SQLAlchemy==1.4.18
tabulate==0.8.10
tmdb3==0.7.2
tmdbsimple==2.8.0
tomli==2.0.1
torrent-tracker-scraper==4.1.0
tracker-scraper==1.1.0
Twisted==21.2.0
two1==3.10.9
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvloop==0.15.2
w3lib==1.22.0
websocket-client==1.4.1
websockets==8.1
yarl==1.6.3
zope.interface==5.4.0
